Refactor (#50)
* Hide the keyboard after clicking login button
* Add DataUtils to contain extensions about bring data to ui
* Fix Usage of colors
* Fix Usages of dimens
