+++ app/src/main/res/layout/bottom_sheet_check_box_list.xml
@@ -17,7 +17,7 @@
         android:layout_height="match_parent"
         android:layout_gravity="bottom"
         android:layoutDirection="rtl"
-        android:minHeight="500dp"
+        android:minHeight="@dimen/min_bottom_sheet_height"
         app:behavior_hideable="true"
         app:layout_behavior="android.support.design.widget.BottomSheetBehavior">
 
@@ -36,7 +36,7 @@
             android:id="@+id/et_search"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
+            android:layout_marginHorizontal="@dimen/spacing_4x"
             android:layout_marginTop="@dimen/spacing_8x"
             android:background="@drawable/bg_edit_text"
             android:drawableStart="@drawable/ic_search"
@@ -49,7 +49,7 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_item_list"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
+            android:layout_height="@dimen/spacing_zero"
             android:layout_marginTop="@dimen/spacing_4x"
             android:clipToPadding="false"
             android:paddingHorizontal="@dimen/spacing_4x"
