+++ app/src/main/res/layout/fragment_student_list.xml
@@ -36,7 +36,7 @@
             android:id="@+id/et_search"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
+            android:layout_marginHorizontal="@dimen/spacing_4x"
             android:layout_marginTop="@dimen/spacing_4x"
             android:background="@drawable/bg_edit_text"
             android:drawableStart="@drawable/ic_search"
@@ -85,7 +85,7 @@
         <View
             android:id="@+id/v_divider"
             android:layout_width="match_parent"
-            android:layout_height="1dp"
+            android:layout_height="@dimen/divider_height"
             android:layout_marginTop="@dimen/spacing_4x"
             android:background="@color/color_border_grey300"
             app:layout_constraintTop_toBottomOf="@id/inc_show_absents" />
@@ -107,7 +107,7 @@
                 tools:text="ریاضی ۲" />
 
             <View
-                android:layout_width="0.5dp"
+                android:layout_width="@dimen/divider_height_half"
                 android:layout_height="match_parent"
                 android:layout_marginHorizontal="@dimen/spacing_2x"
                 android:layout_marginVertical="@dimen/spacing_base"
@@ -126,7 +126,7 @@
                 tools:text="دوشنبه ۷ام فروردین، ساعت ۱۰" />
 
             <View
-                android:layout_width="0.5dp"
+                android:layout_width="@dimen/divider_height_half"
                 android:layout_height="match_parent"
                 android:layout_marginHorizontal="@dimen/spacing_2x"
                 android:layout_marginVertical="@dimen/spacing_base"
@@ -149,7 +149,7 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_student_list"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
+            android:layout_height="@dimen/spacing_zero"
             android:layout_marginTop="@dimen/spacing_4x"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             app:layout_constraintBottom_toBottomOf="parent"
