+++ app/src/main/res/layout/item_check_box.xml
@@ -31,9 +31,9 @@
             android:layout_height="wrap_content"
             android:buttonTint="@color/color_primary"
             android:checked="@{item.checked}"
-            android:minWidth="0dp"
-            android:minHeight="0dp"
-            android:padding="0dp"
+            android:minWidth="@dimen/spacing_zero"
+            android:minHeight="@dimen/spacing_zero"
+            android:padding="@dimen/spacing_zero"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
