+++ app/src/main/res/layout/tag_exam_status.xml
@@ -13,7 +13,7 @@
     <com.google.android.material.textview.MaterialTextView
         setBackgroundDrawableById="@{tag.backgroundDrawable}"
         setTextColorById="@{tag.textColor}"
-        android:layout_width="0dp"
+        android:layout_width="@dimen/spacing_zero"
         android:layout_height="wrap_content"
         android:maxLines="1"
         android:minWidth="@dimen/spacing_20x"
